* Getting rid of Word2011 warnings (disabled in clang) * Getting rid of missing setter warnings for properties in categories (which can't have setters...) - set properties to @dynamic for those categories * setting #pragma unused for variables Xcode things aren't used, but are - ones where we have to fire off messages to invoke things but have no use for the return * Commented out obviously unused items (didn't want to risk removal right now) * Tried first pass at cleaning up nullability warnings by setting nullable / nonnull for some methods / properties * Fixed bug with table preview - was not previously monitoring for changes in ColumnFilter.Enabled - was duplicating RowFilter.Enabled listener (d'oh!). Fixed. * Imported some .h files in places to suppress delegate protocol warnings ("not known!") * Trying to see if we can silence the 10.10 compatibility warnings re: multiline labels. Look at SamplesDownloadProgressViewController.m for example. We're using viewWillLayout to manually set label width (labels need to be changed from automatic to explicit in IB)

+++ StatTag/View Controllers/ScintillaHelpers/SCHelpers.m
@@ -12,7 +12,7 @@
 
 @implementation SCHelpers
 
-+(int)Clamp:(NSInteger)value min:(int)min max:(int) max
++(int)Clamp:(int)value min:(int)min max:(int) max
 {
   if (value < min)
     return min;
@@ -20,7 +20,7 @@
   if (value > max)// && max > min)//eww changed to add && max > min
     return max;
   
-  return value;
+  return (int)value;
 }
